Replace `u16` with `NonZeroU16`
This will remove one 'illegal' state from `Roman` - `Roman(0)`.
Also this allow compiler to make niche-optimizations.

+++ src/error.rs
@@ -10,7 +10,7 @@ pub enum Error {
     InvalidDigit(u8),
 
     /// Value out of range.
-    OutOfRange(i32),
+    OutOfRange(u16),
 }
 
 impl Display for Error {
